New API to work with Gremlin  queries.
