+++ core/src/main/java/com/jetbrains/youtrack/db/api/gremlin/YTDBGraph.java
@@ -3,6 +3,7 @@ package com.jetbrains.youtrack.db.api.gremlin;
 import com.jetbrains.youtrack.db.api.DatabaseSession;
 import com.jetbrains.youtrack.db.internal.core.db.DatabaseSessionEmbedded;
 import com.jetbrains.youtrack.db.internal.core.gremlin.YTDBSingleThreadGraph;
+import org.apache.commons.configuration2.BaseConfiguration;
 import org.apache.tinkerpop.gremlin.structure.Graph;
 import org.apache.tinkerpop.gremlin.structure.util.GraphFactoryClass;
 
@@ -25,7 +26,10 @@ public interface YTDBGraph extends Graph {
       throw new IllegalArgumentException(
           "Passed in database session is not embedded. Only sessions of embedded databases are supported.");
     }
+
+    var config = new BaseConfiguration();
+    config.addProperty(YTDBGraphFactory.CONFIG_YOUTRACK_DB_NAME, session.getDatabaseName());
     return new YTDBSingleThreadGraph(null, (DatabaseSessionEmbedded) session,
-        YTDBGraphFactory.createGraphConfiguration(session));
+        config);
   }
 }
