+++ core/src/main/java/com/jetbrains/youtrack/db/internal/core/tx/FrontendTransactionNoTx.java
@@ -53,6 +53,7 @@ import java.util.Map;
 import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversalSource;
 
 /**
  * No operation transaction.
@@ -302,6 +303,11 @@ public class FrontendTransactionNoTx implements FrontendTransaction {
   @Override
   public ResultSet computeScript(String language, String script, Object... args)
       throws CommandExecutionException, CommandScriptException {
+    throw new UnsupportedOperationException("not supported in no tx mode");
+  }
+
+  @Override
+  public GraphTraversalSource traversal() {
     throw new UnsupportedOperationException("not supported in no tx mode");
   }
